s3UI/s3-team-share/requirements.txt
streamlit
boto3
requests
pyjwt
pycryptodome
